[MODORDERS-1133] - Add support of creating holding in member tenants for Receiving
